cmake: introduce multilib support for MSVC
Note, the multilib configuration exposes a bug in MSVC 2008 in that the
anonymous namespaced functions are not truly file local, the unique namespaced
functions are exported and thus collide with other libx265 static libraries. It
is not known if later MSVC compilers are also affected.

To work around this, we tell the linker to ignore duplicate symbols. This seems
to work but is not a terribly robust long-term solution.